fix: fix form add func, update popup structure
